v1.1.4
return int in hashCode

+++ ticketingsystem/TicketingDS.java
@@ -14,7 +14,7 @@ class TicketWithHash extends Ticket {
     
     @Override
     public int hashCode() {
-        return tid;
+        return (int) tid;
     }
 }
